make next_n run faster
I was creating a list of all numbers up to game.top that haven't been
seen, and shuffling it at each step. I wrote this when game.top didn't grow
exponentially with the score.
Now I've changed it to just pick numbers at random from [1,top] until we
get one we haven't seen. This can get slow when the set of valid numbers
is much smaller than the set of invalid ones, so we use the old method
when num_seen/top >= 0.5..